requirements.txt
click>=8.1.0
motor>=3.3.2
pymongo>=4.6.0
python-dotenv>=1.0.0
pydantic[email]>=2.5.3
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyyaml>=6.0.1
aiohttp>=3.9.0
aioconsole>=0.6.1
rich>=13.7.0
bson>=0.5.10